+++ packages/node/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- `--network-endpoint-config` add optional parameter `limit`
+- `--network-endpoint-config` add optional parameter `pageLimit`
 
 ### Fixed
 - Performance issue caused by making too many RPC requests.