+++ packages/types/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Fixed
-- Performance issue caused by making too many RPC requests.
+- IStellarEndpointConfig add optional parameter `pageLimit`
 
 ## [4.1.0] - 2025-02-05
 ### Changed